NavTracker_python_display/requirements.txt
appdirs==1.4.3
astroid==1.5.2
click==6.7
colorama==0.3.9
Flask==0.12.2
isort==4.2.5
itsdangerous==0.24
Jinja2==2.9.6
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
numpy==1.12.1
packaging==16.8
pandas==0.20.1
pylint==1.7.1
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
pywebview==1.6
selenium==3.4.2
six==1.10.0
tqdm==4.11.2
virtualenv==15.1.0
Werkzeug==0.12.2
<<<<<<< HEAD
astroid==1.5.2
autobahn==17.5.1
bleach==2.0.0
click==6.7
colorama==0.3.9
colosseum==0.1.6
cycler==0.10.0
decorator==4.0.11
entrypoints==0.2.2
Flask==0.12.1
html5lib==0.999999999
ipykernel==4.6.1
ipython==6.0.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
isort==4.2.5
itsdangerous==0.24
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
matplotlib==2.0.2
mccabe==0.6.1
mistune==0.7.4
nbconvert==5.1.1
nbformat==4.3.0
notebook==5.0.0
numpy==1.12.1
pandas==0.20.1
pandocfilters==1.4.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
pyee==3.0.3
Pygments==2.2.0
pylint==1.7.1
pyparsing==2.2.0
PyQt5==5.8.2
pytest-runner==2.11.1
python-dateutil==2.6.0
pythonnet==2.3.0
pythrust==0.7.5
pytz==2017.2
pywebview==1.6
pyzmq==16.0.2
qtconsole==4.3.0
requests==2.13.0
selenium==3.4.1
simplegeneric==0.8.1
sip==4.19.2
six==1.10.0
sofi==0.1.4
testpath==0.3
toga==0.2.8
toga-core==0.2.8
toga-winforms==0.2.8
tornado==4.5.1
tqdm==4.11.2
traitlets==4.3.2
txaio==2.7.1
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.1
widgetsnbextension==2.0.0
=======
>>>>>>> 0b1242b252b618c0a2336d5b80f40ebbf56b780d
wrapt==1.10.10
